eldlogsapp/serializers.py
from rest_framework import serializers

class TripInputSerializer(serializers.Serializer):
    current_location = serializers.CharField()
    pickup_location = serializers.CharField()
    dropoff_location = serializers.CharField()
    hours_used = serializers.FloatField()